Move network features out of ethUtils
