+++ src/network.h
@@ -1,5 +1,4 @@
-#ifndef _NETWORK_H_
-#define _NETWORK_H_
+#pragma once
 
 #include <stdint.h>
 #include <stdbool.h>
@@ -9,4 +8,6 @@ const char *get_network_ticker_from_chain_id(const uint64_t *chain_id);
 
 bool chain_is_ethereum_compatible(const uint64_t *chain_id);
 
-#endif  // _NETWORK_H_
+uint64_t get_tx_chain_id(void);
+
+const char *get_displayable_ticker(const uint64_t *chain_id);
